added new functionality, pagination etc

+++ Backend/Services/Product-service-citymall/src/main/java/com/citymall/dto/ProductResponse.java
@@ -6,7 +6,7 @@ public record ProductResponse(
         Integer id,
         String name,
         String description,
-        double availableQuantity,
+        Integer availableQuantity,
         BigDecimal price,
         Integer categoryId,
         String categoryName,
