+++ Backend/Services/Product-service-citymall/src/main/resources/db/migration/V1__init_database.sql
@@ -10,7 +10,7 @@ create table if not exists product
 (
     id                 integer          not null
         primary key,
-    available_quantity double precision not null,
+    available_quantity integer not null,
     description        varchar(255),
     name               varchar(255),
     price              numeric(38, 2),
